Finish steps DB and add google map

+++ app/src/main/java/com/example/calorietrackerapp/controller/asynctask/CreateFoodAsyncTask.java
@@ -2,8 +2,8 @@ package com.example.calorietrackerapp.controller.asynctask;
 
 import android.os.AsyncTask;
 
-import com.example.calorietrackerapp.restclient.entity.Food;
-import com.example.calorietrackerapp.restclient.service.FoodService;
+import com.example.calorietrackerapp.model.entity.Food;
+import com.example.calorietrackerapp.model.service.FoodService;
 
 import java.util.UUID;
 
@@ -17,7 +17,7 @@ public class CreateFoodAsyncTask extends AsyncTask<String, Void, Void> {
         food.setFoodName(params[0]);
         food.setCategory(params[1]);
         food.setServiceUnit(params[2]);
-        food.setFat(Integer.parseInt(params[3]));
+        food.setFat(Double.parseDouble(params[3]));
         food.setCalorieAmount(Double.parseDouble(params[4]));
         service.createFood(food);
         return null;
